Browser compatibility detection
Logics now looks at the user agent string of a browser, and alerts the user if their browser is incompatible. I'm planning on doing browser compatibility updates in the future, but for now this will do. Logics does not run on mobile devices and it probably never will. I think it's really important that the users at least know why it is not working.

Also code cleanup and inzoom limit.